refactor truncated distributions into a separate file
